Misc code formatting
