+++ lib/core/application.js
@@ -2,9 +2,7 @@
 
 const Koa = require('koa');
 const { Context } = require('@loopback/context');
-
 const Server = require('./server');
-const middlewares = require('./middlewares');
 const BootApplication = require('../boot/boot.application');
 
 class Application extends Context {
@@ -42,7 +40,7 @@ class Application extends Context {
   /**
    * Retrive server of the application
    * 
-   * @return { @luharjs/server } Node Server
+   * @return { server } Node Server
    */
   get server() {
     return this._server;
@@ -66,7 +64,7 @@ class Application extends Context {
    */
   _modifyListener() {
     
-    this._listener.use(middlewares(this.find('middleware.*')));
+    this._listener.use(this.composeMiddlewares());
 
     // Koa callback for starting application
     this.requestListener = this._listener;
@@ -79,6 +77,38 @@ class Application extends Context {
 
     // Starting server
     await this._server.start();
+  }
+
+  /**
+   * Compose middleware is function to combining multiple
+   * middlewares to sequencial execution.
+   */
+  composeMiddlewares() {
+
+
+    let middlewareFn = [];
+
+    for(let middleware of this.find('middleware.*')) {
+      middlewareFn.push(middleware.getValue());
+    }
+
+    return (context, next) => {
+      // last called middleware #
+      let index = -1
+      return execute(0)
+      function execute (i) {
+        if (i <= index) return Promise.reject(new Error('next() called multiple times'))
+        index = i
+        let fn = middlewareFn[i]
+        if (i === middlewareFn.length) fn = next
+        if (!fn) return Promise.resolve()
+        try {
+          return Promise.resolve(fn(context, execute.bind(null, i + 1)));
+        } catch (err) {
+          return Promise.reject(err)
+        }
+      }
+    };
   }
 }
