+++ lib/core/request-handler.js
@@ -3,7 +3,6 @@
 const { RequestContext } = require('../context');
 const { Context } = require('@loopback/context');
 const RoutingTable = require('../routes/routing-table');
-let _ = require('lodash');
 const async = require('async');
 const { bindHeadersToContext } = require('../context');
