+++ lib/core/server.js
@@ -2,7 +2,6 @@
 
 const http  = require('http');
 const https = require('https');
-// const pEvent = require('p-event');
 const { Context } = require('@loopback/context');
 const RequestHandler = require('./request-handler');
