+++ lib/routes/routing-table.js
@@ -1,6 +1,6 @@
 "use strict";
 
-let _ = require('lodash');
+const _ = require('lodash');
 const path = require('./utils/path');
 const {openAPIParameterResolver} = require('./openapi');
 const debug = require('debug')('compositjs:routing-table');
