+++ lib/services/rest-service.controller.js
@@ -12,16 +12,6 @@ const Confidence = require('confidence');
 const CONTEXT_PREFIX = '$.';
 
 /**
- * Return value from environment variable
- * 
- * @param {string} key 
- * @returns value from env
- */
-const getValueFromEnv = (key) => {
-  return process.env[key] || undefined;
-};
-
-/**
  * This function returns resolved paramters which defined in service configuration.
  * 
  * @example Accessing parameter value from context object
